additional comments and re-org functions and files. use if(len(string) == 0) instead of if(string == "").
